feat: 动态创建/修改/删除xxl-job定时任务 && 指定人群文件定时发送消息模板

+++ mpp-support/src/main/java/com/boat/mpp/support/domain/MessageTemplate.java
@@ -31,7 +31,7 @@ public class MessageTemplate implements Serializable {
     private Integer msgStatus;
 
     @Column(name = "cron_task_id")
-    private Long cronTaskId;
+    private Integer cronTaskId;
 
     @Column(name = "cron_crowd_path", length = 500)
     private String cronCrowdPath;
